[form-factors] Split analytic-b-to-v-lcsr.cc file
