+++ eos/form-factors/Makefile.am
@@ -9,7 +9,8 @@ libeosformfactors_la_SOURCES = \
 	analytic-b-to-pi.hh analytic-b-to-pi.cc \
 	analytic-b-to-pi-pi.hh analytic-b-to-pi-pi.cc \
 	analytic-b-to-p-lcsr.hh analytic-b-to-p-lcsr-impl.hh analytic-b-to-p-lcsr.cc \
-	analytic-b-to-v-lcsr.hh analytic-b-to-v-lcsr-impl.hh analytic-b-to-v-lcsr.cc \
+	analytic-b-to-v-lcsr.hh analytic-b-to-v-lcsr-impl.hh \
+	analytic-b-to-rho-lcsr.cc analytic-b-to-kstar-lcsr.cc analytic-b-to-dstar-lcsr.cc \
 	baryonic.cc baryonic.hh baryonic-impl.hh \
 	b-lcdas.cc b-lcdas.hh \
 	form-factor-adapter.hh \
